src/modules/appointments/services/CreateAppointment.spec.ts
import 'reflect-metadata';

import AppError from '@shared/errors/AppError';
import CreateAppointmentService from './CreateAppointment';
import FakeAppointmentsRepository from '../repositories/fakes/FakeAppointmentsRepository';

describe('CreateAppointment', () => {
    it('should be able to create a new appointment', async () => {
        const fakeAppointmentsRepository = new FakeAppointmentsRepository();
        const createAppointment = new CreateAppointmentService(
            fakeAppointmentsRepository,
        );

        const appointment = await createAppointment.execute({
            date: new Date(),
            providerId: '321331213',
        });

        expect(appointment).toHaveProperty('id');
        expect(appointment.providerId).toBe('321331213');
    });

    it('should not be able to create two appointments on the same time', async () => {
        const fakeAppointmentsRepository = new FakeAppointmentsRepository();
        const createAppointment = new CreateAppointmentService(
            fakeAppointmentsRepository,
        );

        const appointmentDate = new Date(2020, 4, 10, 11);

        await createAppointment.execute({
            date: appointmentDate,
            providerId: '321331213',
        });

        expect(
            createAppointment.execute({
                date: appointmentDate,
                providerId: '321331213',
            }),
        ).rejects.toBeInstanceOf(AppError);
    });
});
